[2020-16] complete part B
Use a greedy optimization to match a field to a rule (spec)
only when there is only a single rule that meets that field index.
When there are 2 or more rules that match, skip that rule for now.
This way, the time complexity is linear with regards
to the number of specs, and not factorial.